requirement.txt
Django
django-environ
psycopg2-binary
djangorestframework
pandas
openpyxl
pydantic==1.10.2
dependency-injector